Move project sources under RewardFlow_API, reorganizing existing files/namespaces and Obsidian docs. Add a new Rewards feature slice (entities, DbContext, configurations, endpoints, calculators, factories and session-reward logic) plus related DTOs and interfaces. Introduce Common utilities: IEmployeeLookupService and DTOs, Optional<T> struct, EntityTracker<T>, and enMode enum. Add tests (RewardFlow_Test), solution/project files, many architecture decision records and docs, and update workspace/Obsidian settings. Remove an obsolete top-level Doc/.obsidian/app.json and the old HTTP file.

+++ src/RewardFlow.API/Employees/CreateEmployee/CreateEmployee.cs
@@ -14,7 +14,7 @@ public static class CreateEmployee
         string Name,
         string? NationalNumber,
         string? AccountNumber,
-        float? Salary,
+        decimal? Salary,
         int? FacultyId,
         int? DepartmentId,
         byte? JobTitle,
